+++ example/package.json
@@ -28,7 +28,7 @@
     "react-native-localize": "^2.0.1",
     "react-native-paper": "^4.5.0",
     "react-native-vector-icons": "https://github.com/RichardLindhout/react-native-vector-icons",
-    "react-native-web": "^0.14.8",
+    "react-native-web": "^0.15.0",
     "react-scripts": "3.4.3"
   },
   "devDependencies": {
